Copy ids and attributes to markdown card title
